divide editor controller from paragraph controller
