+++ zeppelin-web/app/index.html
@@ -149,6 +149,7 @@ limitations under the License.
     <script src="scripts/js/directives/resizable.js"></script>
     <script src="scripts/js/directives/ngdelete.js"></script>
     <script src="scripts/js/controllers/paragraph.js"></script>
+    <script src="scripts/js/controllers/paragraph_editor.js"></script>
     <script src="scripts/js/controllers/nav.js"></script>
     <script src="scripts/js/events.js"></script>
     <!-- endbuild -->
