+++ zeppelin-web/app/views/paragraph.html
@@ -34,7 +34,8 @@ limitations under the License.
 
   <div>
     <div ng-show="!paragraph.config.editorHide">
-      <div id="{{paragraph.id}}_editor"
+      <div ng-controller="ParagraphEditorCtrl"
+           id="{{$parent.paragraph.id}}_editor"
            style="opacity: 1;"
            class="editor"
            ui-ace="{
@@ -42,7 +43,7 @@ limitations under the License.
                      onChange: aceChanged,
                      require : ['ace/ext/language_tools']
                    }"
-           ng-model="paragraph.text"
+           ng-model="$parent.paragraph.text"
            ng-class="{'disable': paragraph.status == 'RUNNING' || paragraph.status == 'PENDING' }">
       </div>
     </div>
